+++ src/MoreIcons.cpp
@@ -1,5 +1,4 @@
 #include "MoreIcons.hpp"
-#include "hooks/GJRobotSprite.hpp"
 
 using namespace geode::prelude;
 
@@ -58,8 +57,11 @@ std::vector<std::filesystem::directory_entry> MoreIcons::naturalSort(const std::
     return entries;
 }
 
-void MoreIcons::loadIcons(const std::filesystem::path& path, std::vector<std::string>& list, std::unordered_map<std::string, std::string>& textures) {
-    log::info("Loading {}s", path.filename().string());
+void MoreIcons::loadIcons(const std::filesystem::path& path, std::vector<std::string>& list, std::unordered_map<std::string, std::string>& textures, IconType type) {
+    auto folder = path.filename().string();
+    log::info("Loading {}s", folder);
+    if (LOADING_LAYER)
+        static_cast<CCLabelBMFont*>(LOADING_LAYER->getChildByID("geode-small-label-2"))->setString(fmt::format("More Icons: Loading {}s", folder).c_str());
     if (!std::filesystem::exists(path)) {
         std::filesystem::create_directories(path);
         return;
@@ -115,8 +117,9 @@ void MoreIcons::loadIcons(const std::filesystem::path& path, std::vector<std::st
 
         auto dict = CCDictionary::createWithContentsOfFileThreadSafe(plistPath.c_str());
         auto frames = CCDictionary::create();
+        auto name = std::filesystem::path(noGraphicPlistPath).stem().string();
         for (auto [frameName, frame] : CCDictionaryExt<std::string, CCDictionary*>(static_cast<CCDictionary*>(dict->objectForKey("frames")))) {
-            frames->setObject(frame, ""_spr + frameName);
+            frames->setObject(frame, getFrameName(frameName, name, type));
         }
         dict->setObject(frames, "frames");
         auto metadata = static_cast<CCDictionary*>(dict->objectForKey("metadata"));
@@ -127,7 +130,6 @@ void MoreIcons::loadIcons(const std::filesystem::path& path, std::vector<std::st
         }
         _addSpriteFramesWithDictionary(dict, textureCache->addImage(fullTexturePath.c_str(), false));
         dict->release();
-        auto name = std::filesystem::path(noGraphicPlistPath).stem().string();
         list.push_back(name);
         textures.emplace(name, fullTexturePath);
     }
@@ -139,11 +141,11 @@ void MoreIcons::changeSimplePlayer(SimplePlayer* player, const std::string& file
     if (!player) return;
 
     if (iconType == IconType::Robot) {
-        MIRobotSprite::useCustomRobot(player->m_robotSprite, file);
+        useCustomRobot(player->m_robotSprite, file);
         return;
     }
     else if (iconType == IconType::Spider) {
-        MIRobotSprite::useCustomSpider(player->m_spiderSprite, file);
+        useCustomSpider(player->m_spiderSprite, file);
         return;
     }
 
@@ -197,5 +199,48 @@ void MoreIcons::changeSimplePlayer(SimplePlayer* player, const std::string& file
     if (extraVisible) {
         player->m_detailSprite->setDisplayFrame(extraFrame);
         player->m_detailSprite->setPosition(firstCenter);
+    }
+}
+
+void MoreIcons::useCustomSprite(GJRobotSprite* robot, const std::string& file) {
+    auto spriteParts = robot->m_paSprite->m_spriteParts;
+    auto spriteFrameCache = CCSpriteFrameCache::get();
+    for (int i = 0; i < spriteParts->count(); i++) {
+        auto spritePart = static_cast<CCSpritePart*>(spriteParts->objectAtIndex(i));
+        auto tag = spritePart->getTag();
+
+        auto spriteFrame = fmt::format("{}_{:02}_001.png"_spr, file, tag);
+        auto sprite2Frame = fmt::format("{}_{:02}_2_001.png"_spr, file, tag);
+        auto spriteExtraFrame = fmt::format("{}_{:02}_extra_001.png"_spr, file, tag);
+        auto spriteGlowFrame = fmt::format("{}_{:02}_glow_001.png"_spr, file, tag);
+
+        spritePart->setBatchNode(nullptr);
+        spritePart->setDisplayFrame(spriteFrameCache->spriteFrameByName(spriteFrame.c_str()));
+        if (auto secondSprite = static_cast<CCSprite*>(robot->m_secondArray->objectAtIndex(i))) {
+            secondSprite->setBatchNode(nullptr);
+            secondSprite->setDisplayFrame(spriteFrameCache->spriteFrameByName(sprite2Frame.c_str()));
+            secondSprite->setPosition(spritePart->getContentSize() / 2);
+        }
+
+        auto glowChild = static_cast<CCSprite*>(robot->m_glowSprite->getChildren()->objectAtIndex(i));
+        glowChild->setBatchNode(nullptr);
+        glowChild->setDisplayFrame(spriteFrameCache->spriteFrameByName(spriteGlowFrame.c_str()));
+
+        if (spritePart == robot->m_headSprite) {
+            auto extraFrame = spriteFrameCache->spriteFrameByName(spriteExtraFrame.c_str());
+            auto hasExtra = MoreIcons::doesExist(extraFrame);
+            if (hasExtra) {
+                if (robot->m_extraSprite) {
+                    robot->m_extraSprite->setBatchNode(nullptr);
+                    robot->m_extraSprite->setDisplayFrame(extraFrame);
+                }
+                else {
+                    robot->m_extraSprite = CCSprite::createWithSpriteFrame(extraFrame);
+                    robot->m_headSprite->addChild(robot->m_extraSprite, 2);
+                }
+                robot->m_extraSprite->setPosition(spritePart->getContentSize() / 2);
+            }
+            robot->m_extraSprite->setVisible(hasExtra);
+        }
     }
 }