+++ src/hooks/GJRobotSprite.hpp
@@ -1,8 +1,0 @@
-#include <Geode/modify/GJRobotSprite.hpp>
-class MIRobotSprite {
-public:
-    static void useCustom(GJRobotSprite*, std::string);
-    static void useCustomSprite(GJRobotSprite*, bool);
-    static void useCustomRobot(GJRobotSprite*, std::string);
-    static void useCustomSpider(GJRobotSprite*, std::string);
-};